Moved Resources views to separate files
