+++ presqt/api_v1/utilities/io/write_file.py
@@ -15,7 +15,9 @@ def write_file(file_path, contents, is_json=False):
     is_json : boolean
         Flag representing if the file should be JSON.
     """
+    # Create Directory
     os.makedirs(os.path.dirname(file_path), exist_ok=True)
+    # Write file
     if is_json:
         with open(file_path, 'w') as outfile:
             json.dump(contents, outfile)
