+++ presqt/api_v1/views/resource/__init__.py
@@ -1,0 +1,9 @@
+from presqt.api_v1.views.resource.resource import Resource
+from presqt.api_v1.views.resource.resource_collection import ResourceCollection
+from presqt.api_v1.views.resource.resource_download import PrepareDownload
+
+__all__ = [
+    Resource,
+    ResourceCollection,
+    PrepareDownload
+]
